Update and rename .py to Geeks And The String.py
