--- January 2023/23 January 2023/.py
@@ -1,67 +1,0 @@
-#User function Template for python3
-
-class Solution:
-    
-    #Function to remove pair of duplicates from given string using Stack.
-    def removePair(self,s):
-        # code here
-          st=[]
-
-        l=list(s)
-
-        for i in range(len(l)):
-
-            if len(st)==0:
-
-                st.append(l[i])
-
-            elif st[-1]==s[i]:
-
-                st.pop(-1)
-
-
-
-            else:
-
-                st.append(s[i])
-
-        if len(st)==0:
-
-            return -1
-
-        else:
-
-
-
-            return "".join(st)
-        #User function Template for python3
-
-#{ 
- # Driver Code Starts
-#Initial Template for Python 3
-
-import atexit
-import io
-import sys
-
-#Contributed by : Nagendra Jha
-
-
-_INPUT_LINES = sys.stdin.read().splitlines()
-input = iter(_INPUT_LINES).__next__
-_OUTPUT_BUFFER = io.StringIO()
-sys.stdout = _OUTPUT_BUFFER
-
-@atexit.register
-
-def write():
-    sys.__stdout__.write(_OUTPUT_BUFFER.getvalue())
-    
-
-
-if __name__ == '__main__':
-    test_cases = int(input())
-    for cases in range(test_cases) :
-        obj = Solution()
-        print(obj.removePair(str(input())))
-# } Driver Code Ends
